make direction (turn) alerts based on speed so we consitently get an alert "x seconds" away from the corner
a fixed distance was not good enough because if the user speeds up in some sections they want the alert to be triggered with enough time to react.
If running the distance would have to be larger than if walking.
With a time based approach it means the number can be fixed and the setting does not need to be cahnged when switching activities.