remove progress bar for "clear stored tiles"
It looks like onUpdate/unLayout are not called on any views pushed to the view stack whilst a confirmation is active
I have no way to defer the result of the confirmation, since I cannot start timers in a datafield
So the storage cache has to be cleared when the user clicks the confirm button on "clear cached tiles"
This is not ideal since it takes some time to complete, and the user might think the ui is frozen. It should hopefully be a rare operation though.
Maybe I can revisit this by providing my own 'confirmation' ui too, but that would add even more code (and memory is a valuable resource)