Fix issue with wrong direction alerts
We wer using the distance from user to next point instead of the distance from

if the users moves off the track but is still 'on track' then the distance increases through the hypotenuse, but what we actually care about is have they gone backwards on the stright line connecting the 2 points. not have they gone sideways.